stop scrolling at bottom of node
git-svn-id: svn://svn.sv.gnu.org/texinfo/trunk@5486 39fee189-59d7-47db-b5d4-205258b72aed

+++ info/window.c
@@ -587,7 +587,7 @@ window_toggle_wrap (WINDOW *window)
   if (window != the_echo_area)
     {
       char **old_starts;
-      size_t *old_xlat;
+      long *old_xlat;
       int old_lines, old_pagetop;
 
       old_starts = window->line_starts;
